+++ CHANGELOG.md
@@ -10,6 +10,7 @@ Changelog
 - Changed `dukt\social\linkedin\loginproviders\Linkedin::getOauthProvider()` method’s visibility to `public`.
 - Renamed `dukt\social\linkedin\loginproviders\Linkedin::getDefaultScope()` to `dukt\social\linkedin\loginproviders\Linkedin::getDefaultOauthScope()`.
 - Removed `dukt\social\linkedin\loginproviders\Linkedin::getProfile()`.
+- Removed `dukt\social\linkedin\loginproviders\Linkedin::getRemoteProfile()`.
 
 ## 3.0.2 - 2018-05-18